Main Code v1 and MPA v1
I have added a new main file to call the new version of the MPA. In this file, we find the best subset of antennas for each edge of the graph. Then, the MPA is applied mixing up the signal from the antennas of these optimal subsets. The criteria is the SIR (signal-to-interference ratio). Use TB_ESGA_2020_v1 for the new simulations.